move BubblingSample to separate class
